Refactored code for utils
